Update import after propTypes file rename
